requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Redis
redis==5.2.1
aioredis==2.0.1

# Elasticsearch
elasticsearch[async]==8.17.1

# Message queue (Celery + RabbitMQ)
celery[redis]==5.4.0
kombu==5.4.2

# Auth & validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email-validator]==2.10.3
pydantic-settings==2.6.1

# Monitoring (Prometheus)
prometheus-client==0.21.0

# HTTP client
httpx==0.28.1

# Testing (TDD/BDD)
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-bdd==7.0.1
aiosqlite==0.20.0
httpx==0.28.1

# Dev & quality
black==24.10.0
ruff==0.8.2
